+++ CHANGELOG.md
@@ -31,6 +31,8 @@ and a dynamic pool for those that are larger.
 
 - GCC 4.9 build to Travis CI.
 
+- Added a new IOManager that stores logging and replay output to files.
+
 ### Changed
 
 - Replay uses JSON format for its I/O.